api/requirements.txt
# AWS provides these in the runtime, so to keep package size down at least for now don't bundle them
# boto3==1.19.2
# botocore==1.22.2
backports.zoneinfo
psycopg[binary]==3.0.1